requirements.txt
# To activate this run
# pip install -r requirements.txt
#
# or go to Python Packages
# to install all the required packages by hand
#
#
Django==4.1.6
djangorestframework==3.14.0
django-extensions==3.2.1
django-debug-toolbar~=3.8.1
psycopg2
django-autoslug~=1.9.8
python-decouple